requirements.txt
asgiref==3.8.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
crispy-bootstrap5==2024.10
cryptography==45.0.3
Django==5.1.5
django-allauth==65.8.1
django-ckeditor==6.7.2
django-crispy-forms==2.3
django-jazzmin==3.0.1
django-js-asset==3.0.1
django-parler==2.3
django-rosetta==0.10.1
idna==3.10
mysqlclient==2.2.7
oauthlib==3.2.2
pillow==11.1.0
polib==1.2.0
pycparser==2.22
PyJWT==2.10.1
PyMySQL==1.1.1
requests==2.32.3
requests-oauthlib==2.0.0
setuptools==80.9.0
sqlparse==0.5.3
urllib3==2.3.0
wheel==0.45.1
